Found the problem but not the solution
Below is the issue with the Summary code I want to write.

The problem seems to fit well into the monadic context. There is a
thing `a` that carries with it a tree of summaries of past things.

This can be expressed as

```
data Cold s   = Cold s [Cold s]
data Hot  s d = Hot  d [Cold s]

instance Monad (Hot s) where

  return x = Hot x []

  (Hot x1 cs) >>= f = case f x1 of
    Hot x2 [] -> Hot x2 [Cold (summary x2) xs]
```

Here the return function simply loads the new data into the `Hot`
structure (which holds the data that is ready for use). The bind
function summarizes the data in `Hot`, adds the summary to the `Cold`
history tree, and transforms the `Hot` data into the new `Hot` data.

The basic problem here, is that the `summary` function needs to have the
type signature `a -> s`, where `s` is a rigid type variable set by the
instance. This summary function needs to work for every type `a`.

I think I need some fancy type voodoo.