Add Report system
This merges the concepts of Error, Warn, and Note. Warnings and notes
are passed along with the data. The comment on results, but are not
fatal. Errors are terminal, pass on no value. Report has the structure:

```
data Report e a = Pass a e e | Fail e e e
```

Where e is a monoid (a restriction added in the Monad, Functor, and
Applicative instances). The e's in Pass correspond to Warning and Note,
the first e in Fail, is Error. The differences between this and Either
is that additional information can be passed, and is accumulated as it
goes. In the applicative instance, all error messages are merged. For
example in `foo <$> x <*> y <*> z <*>`, the error messages from x, y,
and z will all be merged together, in contrast to Either, where only the
first error is passed.