Get Report working everywhere
Now the whole program compiles and the Main runs a sample program.

This is working beautifully now, I can handle both monadic error and
applicative error. In monadic error, the first error propagate
throughout the program, all future operations are bypassed. Which is
exactly how error in a sequence should be handled.

However, parallel error should not propagate, but rather merge. If a
function takes several variables, and multiple errors occur, it is
useful to be able to report all the errors.

For example, say we have

```
foo :: a -> b -> Report e d
bar :: a -> Report e b
baz :: a -> Report e b

foo <$> (bar x >>= baz) <*> bar y
```

`baz` and `bar y` can be run in parallel, and both might return an
error. If Either were used, only the error of `baz` would be reported.
Report returns both errors. The warnings and notes it passes similarly,
except that the computed value is preserved.